Organized and split result labels
I split the result label into prediction and certainty labels.

I also packed the objects top-bottom instead of placing them to help with centering and organization.

I removed label appearance before image selection and changed the coloring of the select image button (this might be a problem for different OS or Themes).